// src/server/api/models/investing/models/FixedDepositTransactionRequest.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

/**
 * A JSON blob representing a fixed deposit transaction request.
 */
export type FixedDepositTransactionRequest = {
    /**
     * The rollover rule
     */
    rolloverRule: FixedDepositTransactionRequest.rolloverRule;
    /**
     * The investment account ID
     */
    accountId: string;
    /**
     * The product ID
     */
    productId: string;
    /**
     * Notes related to the record
     */
    notes?: string;
    /**
     * The transaction currency
     */
    currency: string;
    /**
     * The face value amount
     */
    principal: number;
    /**
     * The start date
     */
    startDate: string;
    /**
     * The tenor
     */
    tenor: number;
    /**
     * The interest rate
     */
    interestRate?: number;
    /**
     * The reference transaction code. This is usually a code generated by an upstream or used for uploading records
     */
    refCode?: string;
    /**
     * The position transfer date is used to indicate that the transaction was migrated from another system and accruals before that date should not be considered.
     */
    positionTransferDate?: string;
    /**
     * The channel through which the transaction was submitted
     */
    channel?: FixedDepositTransactionRequest.channel;
};

export namespace FixedDepositTransactionRequest {

    /**
     * The rollover rule
     */
    export enum rolloverRule {
        PRINCIPAL = 'PRINCIPAL',
        PRINCIPAL_INTEREST = 'PRINCIPAL_INTEREST',
        INTEREST = 'INTEREST',
        NONE = 'NONE',
    }

    /**
     * The channel through which the transaction was submitted
     */
    export enum channel {
        NA = 'NA',
        WEB = 'WEB',
        MOBILE = 'MOBILE',
        USSD = 'USSD',
        ATM = 'ATM',
        POS = 'POS',
        BRANCH = 'BRANCH',
        EXEC_REPORT = 'EXEC_REPORT',
        FILE_UPLOAD = 'FILE_UPLOAD',
        API = 'API',
        EMAIL = 'EMAIL',
        SMS = 'SMS',
        IVR = 'IVR',
        KIOSK = 'KIOSK',
        CALL_CENTER = 'CALL_CENTER',
        OTHER = 'OTHER',
    }


}